Move TextRangeWithAddon component to src/components/elements/propertiesTab
