+++ app/packs/src/components/elements/samples/propertiesTab/TextRangeWithAddon.js
